ADD Clean and split code into a folder -- better code review
 Committer: apayet <apayet@localhost.localdomain>

 On branch ap/ws_transmit_basic_data_front_html_js_and_nestjs
 Your branch is up to date with 'origin/ap/ws_transmit_basic_data_front_html_js_and_nestjs'.

 Changes to be committed:
	deleted:    sandbox/front/app/react-ts/src/components/MenuBar.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/AvatarMenu.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/HamburgerMenu.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/IconTitle.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/LargeMenu.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/MenuBar.tsx
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/config/DisplayStyle.ts
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/config/FontConfig.ts
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/config/PagesItem.ts
	new file:   sandbox/front/app/react-ts/src/components/MenuBar/config/SettingsItem.ts
	modified:   sandbox/front/app/react-ts/src/components/ReduxViewTest.tsx
