+++ sandbox/front/app/react-ts/src/components/MenuBar/AvatarMenu.tsx
@@ -1,0 +1,82 @@
+/* eslint-disable max-statements */
+/* eslint-disable max-lines-per-function */
+import	React from "react";
+import	Box from "@mui/material/Box";
+import	Tooltip from "@mui/material/Tooltip";
+import	IconButton from "@mui/material/IconButton";
+import	Avatar from "@mui/material/Avatar";
+import	Menu from "@mui/material/Menu";
+import	MenuItem from "@mui/material/MenuItem";
+import	Typography from "@mui/material/Typography";
+
+import	settings from "./config/SettingsItem";
+
+const	AvatarMenu = () =>
+{
+	const [
+		anchorElUser,
+		setAnchorElUser
+	] = React.useState<null | HTMLElement>(null);
+
+	const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) =>
+	{
+		setAnchorElUser(event.currentTarget);
+	};
+
+	const handleCloseUserMenu = () =>
+	{
+		setAnchorElUser(null);
+	};
+
+	return (
+		<Box sx={{flexGrow: 0}}>
+			<Tooltip title="Open settings">
+				<IconButton
+					onClick={handleOpenUserMenu}
+					sx={{ p: 0}}
+				>
+					<Avatar
+						alt="image profile"
+						src="https://thispersondoesnotexist.com/"
+					/>
+				</IconButton>
+			</Tooltip>
+			<Menu
+				sx={{ mt: "45px" }}
+				id="menu-appbar"
+				anchorEl={anchorElUser}
+				anchorOrigin={
+				{
+					vertical: "top",
+					horizontal: "right",
+				}}
+				keepMounted
+				transformOrigin={
+				{
+					vertical: "top",
+					horizontal: "right",
+				}}
+				open={Boolean(anchorElUser)}
+				onClose={handleCloseUserMenu}
+			>
+			{
+				settings.map((setting) =>
+				{
+					return (
+						<MenuItem
+							key={setting}
+							onClick={handleCloseUserMenu}
+						>
+							<Typography textAlign="center">
+								{setting}
+							</Typography>
+						</MenuItem>
+					);
+				})
+			}
+			</Menu>
+		</Box>
+	);
+};
+
+export default AvatarMenu;
