+++ sandbox/front/app/react-ts/src/components/MenuBar/config/DisplayStyle.ts
@@ -1,0 +1,15 @@
+
+const	displayStyle = {
+	smallHidden:
+	{
+		xs: "none",
+		md: "flex"
+	},
+	mediumHidden:
+	{
+		xs: "flex",
+		md: "none"
+	}
+};
+
+export default displayStyle;
