+++ sandbox/front/app/react-ts/src/components/MenuBar/config/FontConfig.ts
@@ -1,0 +1,14 @@
+
+const fontConfig = {
+	mr: 2,
+	display:
+	{
+		fontFamily: "monospace",
+		fontWeight: 700,
+		letterSpacing: ".3rem",
+		color: "inherit",
+		textDecoration: "none"
+	}
+};
+
+export default fontConfig;
