+++ sandbox/front/app/react-ts/src/components/MenuBar/config/PagesItem.ts
@@ -1,0 +1,8 @@
+
+const pages = [
+	"Home",
+	"Launch Game",
+	"LeaderBoard"
+];
+
+export default pages;
