+++ sandbox/front/app/react-ts/src/components/MenuBar/config/SettingsItem.ts
@@ -1,0 +1,9 @@
+
+const settings = [
+	"Profile",
+	"Account",
+	"Dashboard",
+	"Logout"
+];
+
+export default settings;
